feat: set of components and pages
