+++ src/app/about/about.module.scss
@@ -1,0 +1,87 @@
+
+.article {
+  margin: 2rem 0;
+  display: flex;
+  flex-direction: column-reverse;
+  align-items: flex-start;
+  justify-content: space-between;
+  position: relative;
+  .textContent {
+    padding: 2rem;
+    gap: 1rem;
+    display: flex;
+    flex-direction: column;
+    text-align: left;
+  }
+  &:before {
+    border-radius: var(--border-radius-large);
+    background: var(--background-alpha-70);
+    height: 100%;
+    width: 100%;
+    left: 0;
+    top: 0;
+    position: absolute;
+    z-index: -1;
+    content: " ";
+  }
+  .imageContainer {
+    margin: -2rem auto;
+    img {
+    width: 50vw;
+    height: 50vw;
+    border-radius: 50%;
+    box-shadow: var(--shadow-large);
+    object-fit: cover;
+    object-position: 100% 0;
+    }
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .article {
+    flex-direction: row;
+    &:before {
+      width: 80%;
+    }
+    .textContent {
+      padding: 2rem 4rem;
+      max-width: 50vw;
+    }
+    .imageContainer {
+      margin: 4rem auto;
+      img {
+        width: 100%;
+        height: auto;
+        border-radius: var(--border-radius-large);
+        box-shadow: var(--shadow-large);
+      }
+    }
+  }
+  .button{
+    margin-right: 0;
+  }
+}
+
+/* Enable hover only on non-touch devices */
+@media (hover: hover) and (pointer: fine) {
+
+}
+
+@media (prefers-reduced-motion) {
+
+}
+
+/* Mobile */
+@media (max-width: 700px) {
+
+}
+
+/* Tablet and Smaller Desktop */
+@media (min-width: 701px) and (max-width: 1120px) {
+
+}
+
+@media (prefers-color-scheme: dark) {
+  
+}
+
