+++ src/app/components/button/button.module.scss
@@ -1,0 +1,29 @@
+
+.button {
+    background-color: rgba(0,0,0,0.7);
+    border: 3px solid #039CE3;
+    padding: 1rem 2rem;
+    color: #fff;
+    text-align: center;
+    text-decoration: none;
+    border-radius: var(--border-radius-large);
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    gap: .5rem;
+    font-size: 1em;
+    &:hover {
+        border-color: #7CD6FF;
+        background-color: #000;
+    }
+}
+
+.iconLinkedIn:before{
+    width: 1rem;
+    height: 1rem;
+    flex: 0 0 1rem;
+    display: inline-block;
+    content: " ";
+    background-size: cover;
+    background-image: url(/img/linkedIn.svg);
+}
