+++ src/app/globals.scss
@@ -120,55 +120,13 @@ body {
     transform: scale(1);
   }
   50% {
-    transform: scale(2);
+    transform: scale(1.2);
   }
   100% {
     transform: scale(1);
   }
 }
 
-.wrapper {
-  padding-right: 2rem;
-  padding-left: 2rem;
-}
-
-article {
-  padding: 0;
-  display: flex;
-  flex-direction: column-reverse;
-  align-items: flex-start;
-  justify-content: space-between;
-  position: relative;
-  .textContent {
-    padding: 2rem;
-    gap: 1rem;
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-  }
-  &:before {
-    border-radius: var(--border-radius-large);
-    background: var(--background-alpha-70);
-    height: 100%;
-    width: 100%;
-    left: 0;
-    top: 0;
-    position: absolute;
-    z-index: -1;
-    content: " ";
-  }
-  .imageContainer {
-    margin: -2rem auto;
-    img {
-      width: 50vw;
-      height: 50vw;
-      border-radius: 50%;
-      box-shadow: var(--shadow-large);
-      object-fit: cover;
-      object-position: 100% 0;
-    }
-  }
-}
 
 h1 {
   padding: 2rem 0 2rem 0;
@@ -199,11 +157,11 @@ h3 {
   font-size: var(--font-size-3);
 }
 h4 {
-  paddding: 1rem 0, 1rem 0,;
+  padding: 1rem 0, 1rem 0,;
   font-size: var(--font-size-4);
 }
 h5 {
-  paddding: 1rem 0 1rem 0;
+  padding: 1rem 0 1rem 0;
   font-size: var(--font-size-5);
 }
 p {
@@ -219,156 +177,18 @@ a {
 }
 
 main {
-  padding: 2rem 0; /*use rem*/
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 2rem;
   text-align: center;
-}
-
-nav {
-  display: flex;
-  gap: 0;
-  align-items: center;
-  a {
-    padding: .5rem 1rem .25rem;
-    background-color: transparent;
-    position: relative;
-    text-decoration: none;
-    font-size: var(--font-size-4);
-    &:after {
-      content: " ";
-      position: absolute;
-      height: 1px;
-      background-color: #fff;
-      opacity: 0;
-      width: 0;
-      bottom: 0;
-      left: 50%;
-      transform: translateX(-50%);
-      transition: .3s ease;
-    }
-    &:hover:after,
-    &.active:after {
-      opacity: 1;
-      width: 100%;
-      background-color: #7CD6FF;
-    }
-    &:focus-visible:after {
-      opacity: 1;
-      width: 100%;
-      background-color: #7CD6FF;
-    }
-  }
-}
-
-header {
-  display: flex;
-  flex-direction: column;
-  padding: 4rem 0 2rem;
-  align-items: center;
-  gap: 2rem;
-  justify-content: space-between;
-  img {
-    width: 5rem;
-  }
-}
-
-
-
-.button {
-  background-color: rgba(0,0,0,0.7);
-  border: 3px solid #039CE3;
-  padding: 1rem 2rem;
-  color: #fff;
-  text-align: center;
-  text-decoration: none;
-  border-radius: var(--border-radius-large);
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: .5rem;
-  font-size: 1em;
-  margin: auto;
-  &:hover {
-    border-color: #7CD6FF;
-    background-color: #000;
-  }
-}
-
-.iconLinkedIn:before{
-  width: 1rem;
-  height: 1rem;
-  flex: 0 0 1rem;
-  display: inline-block;
-  content: " ";
-  background-size: cover;
-  background-image: url(/img/linkedIn.svg);
-}
-
-footer {
-  padding: 2rem 0;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-  text-align: center;
-  &:before {
-    width: 100vw;
-    height: 15vh;
-    z-index: -1;
-    left: 0;
-    bottom: 0;
-    position: fixed;
-    content: " ";
-    background: var(--linear-gradient);
-  }
-  a {
-    color: #fff;
-    display: inline-block;
-  }
 }
 
 @media screen and (min-width: 768px) {
   h1{
     text-align: left;
   }
-  .wrapper {
-      padding-right: 7.5rem;
-      padding-left: 7.5rem;
-  }
-  header {
-      flex-direction: row;
-  }
+
   main {
     text-align: left;
-    flex-direction: column;
   }
-  article {
-    flex-direction: row;
-    &:before {
-      width: 80%;
-    }
-    .textContent {
-      padding: 2rem 4rem;
-      max-width: 50vw;
-    }
-    .imageContainer {
-      margin: 4rem auto;
-      img {
-        width: 100%;
-        height: auto;
-        border-radius: var(--border-radius-large);
-        box-shadow: var(--shadow-large);
-      }
-    }
-  }
-  main .button{
-      margin-right: 0;
-  }
-  footer {
-      justify-content: start;
-  }
+
 }
