+++ src/app/home.module.scss
@@ -1,0 +1,46 @@
+.featuredSection {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 0;
+  text-align: center;
+}
+
+@media screen and (max-width: 767px) {
+  .featuredSection {
+    background-color: var(--background-alpha-70);
+    padding: 1rem;
+    border-radius: var(--border-radius-small);
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .featuredSection {
+    text-align: left;
+    flex-direction: column;
+  }
+}
+
+/* Enable hover only on non-touch devices */
+@media (hover: hover) and (pointer: fine) {
+
+}
+
+@media (prefers-reduced-motion) {
+
+}
+
+/* Mobile */
+@media (max-width: 700px) {
+
+}
+
+/* Tablet and Smaller Desktop */
+@media (min-width: 701px) and (max-width: 1120px) {
+
+}
+
+@media (prefers-color-scheme: dark) {
+  
+}
+
