+++ src/app/icons.module.scss
@@ -1,0 +1,9 @@
+.iconLinkedIn:before{
+    width: 1rem;
+    height: 1rem;
+    flex: 0 0 1rem;
+    display: inline-block;
+    content: " ";
+    background-size: cover;
+    background-image: url(/img/linkedIn.svg);
+}
